refactor(text-select): split by what a selection query actually is
1,634 lines holding four distinct jobs: finding a string on a page, resolving the word under a
point, detecting column gutters, and selecting a range. They share a vocabulary — the geometry
types, the character predicates, the line and glyph helpers — which is why they ended up in one
file, but sharing a vocabulary is not the same as being one concern.

  mod.rs      231   NormRect/CharBox/Wrap/TextAnchor/TextSelection/SearchMatch, the character
                    predicates, and the line/glyph helpers every group leans on
  search.rs   124   find_matches, query normalization, result snippets
  word.rs     154   word_at and the wrap handling that makes a line-broken word still one word
  columns.rs  131   gutter detection within the selection's own vertical band
  span.rs     250   text_in_rect and the drag selection

The cross-file surface is five items, and each one says something true about the design.
`search` needs `wrap_of` because a match spanning a line break has to read its wrap off the printed
hyphen — the same fact that stops a lookup asking the dictionary about "well--" (#231). `span`
needs `confine_to_columns` and `glyph_selected` because a drag across a two-column page must not
sweep up the neighbouring column at the same height (#128). Both are imported by name rather than
through the `use super::*` glob, so the dependency is visible at the top of the file that has it.

The public surface is unchanged: `pdf.rs` and `reflow.rs` still write `text_select::word_at(..)`,
not `text_select::word::word_at(..)`. `confine_to_columns` was private before and stays private,
now as `pub(super)`.

The 800-line test suite stays one file rather than being split four ways. It exercises selection
*behaviour*, which is what word, span and columns produce together — splitting it would file each
test under whichever module it happens to call first, which is not where someone would look for it.

All 43 top-level items are byte-identical to before, except `confine_to_columns`, where rustfmt
reflowed the signature onto multiple lines after `pub(super)` made it too long. Gates green.
